+++ lib/tui_tui/key_code.rb
@@ -4,6 +4,8 @@ module TuiTui
   module KeyCode
     ESCAPE = "\e"
     CTRL_C = "\u0003"
+    CTRL_N = "\u000E"
+    CTRL_P = "\u0010"
     BACKSPACE = "\u007F"
   end
 end